chore(react-native): move test fixture scripts to test directory
