+++ .buildkite/react-native-navigation-pipeline.full.yml
@@ -26,7 +26,7 @@ steps:
           - "test/react-native/features/fixtures/generated/react-native-navigation/**/reactnative.apk"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           setup:
             rn_version:
@@ -53,7 +53,7 @@ steps:
           - "test/react-native/features/fixtures/generated/react-native-navigation/**/reactnative.apk"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           setup:
             rn_version:
@@ -82,7 +82,7 @@ steps:
           - "test/react-native/features/fixtures/generated/react-native-navigation/**/reactnative.ipa"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           setup:
             rn_version:
@@ -111,7 +111,7 @@ steps:
       #     - "test/react-native/features/fixtures/generated/react-native-navigation/**/reactnative.ipa"
       #   commands:
       #     - bundle install
-      #     - ./bin/generate-react-native-fixture
+      #     - node test/react-native/scripts/generate-react-native-fixture.js
       #   matrix:
       #     setup:
       #       rn_version:
