+++ .buildkite/react-native-pipeline.full.yml
@@ -22,7 +22,7 @@ steps:
         artifact_paths:
           - "test/react-native/features/fixtures/generated/old-arch/**/reactnative.apk"
         commands:
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           setup:
             reactnative:
@@ -57,7 +57,7 @@ steps:
         artifact_paths:
           - "test/react-native/features/fixtures/generated/new-arch/**/reactnative.apk"
         commands:
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
               - "0.72"
               - "0.74"
@@ -86,7 +86,7 @@ steps:
           - "test/react-native/features/fixtures/generated/native-integration/old-arch/**/reactnative.apk"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           setup:
             reactnative:
@@ -123,7 +123,7 @@ steps:
           - "test/react-native/features/fixtures/generated/native-integration/new-arch/**/reactnative.apk"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
             - "0.72"
             - "0.74"
@@ -151,7 +151,7 @@ steps:
           - "test/react-native/features/fixtures/generated/old-arch/**/output/reactnative.ipa"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           - "0.64"
           - "0.72"
@@ -180,7 +180,7 @@ steps:
           - "test/react-native/features/fixtures/generated/new-arch/**/output/reactnative.ipa"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           - "0.72"
           - "0.74"
@@ -209,7 +209,7 @@ steps:
           - "test/react-native/features/fixtures/generated/native-integration/old-arch/**/output/reactnative.ipa"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           - "0.64"
           - "0.72"
@@ -239,7 +239,7 @@ steps:
           - "test/react-native/features/fixtures/generated/native-integration/new-arch/**/output/reactnative.ipa"
         commands:
           - bundle install
-          - ./bin/generate-react-native-fixture
+          - node test/react-native/scripts/generate-react-native-fixture.js
         matrix:
           - "0.72"
           - "0.74"
