+++ test/react-native/scripts/utils/dependency-utils.js
