+++ test/react-native/scripts/utils/index.js
