+++ test/react-native/scripts/utils/react-native-config.js
